Refactored RequestsUtility to UnityRequestHandler
- UnityRequestHandler It implements IHttpRequestHandler, targeting requests within the Unity platform.
- UnityMyHttpRequestExtension contains methods related to UnityWebRequest conversion methods.
