--- Runtime/Requests/RequestsUtility.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 53344065aefaa614fab45dd2c1eaba72
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
